third_party/WebKit/LayoutTests/http/tests/websocket/handshake-fail-by-404-expected.txt

CONSOLE ERROR: line 15: WebSocket connection to 'ws://localhost:8880/handshake-fail-by-404' failed: Error during WebSocket handshake: Unexpected response code: 404
Connection should fail immediately if the server returns a 404 error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
